databricks-mcp-tool/requirements.txt
databricks-sdk>=0.44.0
anyio>=3.7.1
click>=8.0.0
httpx>=0.24.0
starlette>=0.27.0
uvicorn>=0.23.0
git+https://github.com/modelcontextprotocol/python-sdk.git#egg=mcp
pytest>=7.0.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
